PACKAGES SEND FINALLY.

+++ Project_Client/DataPacket.h
@@ -8,7 +8,7 @@
 struct Header {
 	unsigned char DType : 4;
 	unsigned char Flags : 4;
-	short unsigned int Datasize;
+	unsigned int Datasize;
 };
 struct Tail {
 	unsigned char Seqnum;
@@ -22,10 +22,10 @@ public:
 	DataPkt();
 	~DataPkt();
 	DataPkt(char*);
-	void setHead(unsigned char, unsigned char, short unsigned int);
+	void setHead(unsigned char, unsigned char, unsigned int);
 	void setDType(unsigned char);
 	void setFlags(unsigned char);
-	void setDsize(unsigned short int);
+	void setDsize(unsigned int);
 	void setTBuf(char*, int&);
 	char* getTBuf();
 	int getDType();
